% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_wb.R
\name{decrypt_wb}
\alias{decrypt_wb}
\title{Decrypt a password-protected Excel file}
\usage{
decrypt_wb(path, password, path_out = tempfile(fileext = get_ext(path)))
}
\arguments{
\item{path}{Path to an Excel workbook file (either .xlsx or .xlsb)}

\item{password}{Workbook password}

\item{path_out}{Output path for decrypted workbook}
}
\value{
Path to the decrypted workbook file (same as argument \code{path_out})
}
\description{
Uses the Python library
\href{https://pypi.org/project/msoffcrypto-tool/}{msoffcrypto-tool} to decrypt
password-protected Excel files (either .xlsx or .xlsb), which can then be
read in with e.g. the \link{readxl} or \link[readxlsb:read_xlsb]{readxlsb} R packages.
}
\examples{
path_xlsx <- system.file("extdata", "xltest.xlsx", package = "rpxl")
decrypt_wb(path_xlsx, password = "1234")

}
