% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_rpxl.R
\name{install_rpxl}
\alias{install_rpxl}
\title{Install Python dependencies for the rpxl package}
\usage{
install_rpxl(
  envname = "r-rpxl",
  new_env = identical(envname, "r-rpxl"),
  method = "auto",
  ...
)
}
\arguments{
\item{envname}{Name of Python environment to install within. Defaults to
"r-rpxl".}

\item{new_env}{Logical indicating whether to remove an existing virtual
environment, if it has the name as \code{envname}, before creating the new
virtual environment and installing the required Python packages.}

\item{method}{Installation method. Defaults to "auto" to automatically find a
method that will work in the local environment.}

\item{...}{Additional arguments passed to \link[reticulate:py_install]{reticulate::py_install}}
}
\description{
Install Python dependencies for the rpxl package
}
