% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpxl-package.R
\docType{package}
\name{rpxl-package}
\alias{rpxl}
\alias{rpxl-package}
\title{rpxl: Read Password-Protected Excel Files}
\description{
Uses the Python library msoffcrypto-tool to decrypt password-protected Excel files, which can then be read into R using readxl, readxlsb, or other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epicentre-msf/rpxl}
  \item Report bugs at \url{https://github.com/epicentre-msf/rpxl/issues}
}

}
\author{
\strong{Maintainer}: Patrick Barks \email{patrick.barks@epicentre.msf.org}

}
\keyword{internal}
